refactor: process extensions
